Static cast channel index

+++ media/base/audio_bus.h
@@ -115,8 +115,12 @@ class MEDIA_EXPORT AudioBus {
   // Returns a raw pointer to the requested channel.  Pointer is guaranteed to
   // have a 16-byte alignment.  Warning: Do not rely on having sane (i.e. not
   // inf, nan, or between [-1.0, 1.0]) values in the channel data.
-  float* channel(int channel) { return channel_data_[channel]; }
-  const float* channel(int channel) const { return channel_data_[channel]; }
+  float* channel(int channel) {
+    return channel_data_[static_cast<size_t>(channel)];
+  }
+  const float* channel(int channel) const {
+    return channel_data_[static_cast<size_t>(channel)];
+  }
   void SetChannelData(int channel, float* data);
 
   int channels() const {
